feat: selectable resolution (320x240 / 400x240 / 640x480x1 mono)

+++ include/serial_test.h
@@ -11,8 +11,10 @@
  * and the RP2350 displays it via DVI — no ESP32 required.
  *
  * Frame format (identical to the SPI protocol in frame_config.h):
- *   4 bytes  magic  (0xA5 0xB6 0xC7 0xD8)
- *   153 600  RGB565 pixel data, row-major, top-left origin
+ *   4 bytes            magic  (0xA5 0xB6 0xC7 0xD8)
+ *   FRAME_PIXEL_BYTES  pixel data, row-major, top-left origin.  The pixel
+ *                      width follows the selected mode (WUADVI_PIXEL_BYTES):
+ *                      RGB565 (2 B/px in 320x240 / 400x240).
  *
  * Use the companion Python script  test/send_frame.py  to send frames.
  */
@@ -28,7 +30,8 @@ void serial_test_init(void);
  *   - the 4-byte magic does not match (bad sync; frame discarded), OR
  *   - the pixel data timed out before FRAME_PIXEL_BYTES were received.
  *
- * dvi_fb must point to the PicoDVI framebuffer (dvi_display.getBuffer()).
+ * dvi_fb must point to the PicoDVI framebuffer (dvi_display.getBuffer()),
+ * passed as raw bytes so it works regardless of the selected pixel mode.
  * Pixel data is written directly into it — no intermediate copy needed.
  */
-bool serial_test_poll(uint16_t *dvi_fb);
+bool serial_test_poll(uint8_t *dvi_fb);
